pyproject.toml
[project]
name = "airflow-docker"
version = "0.1.0"
description = ""
authors = [
    {name = "zombielu",email = "ludaiche@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13, <3.14"
dependencies = [
    "apache-airflow (>=3.1.7,<4.0.0)",
    "postgres (>=4.0,<5.0)",
    "pandas (>=3.0.1,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
